+++ apps/api/src/app/benchmark/benchmark.controller.ts
@@ -1,5 +1,6 @@
 import { HasPermission } from '@ghostfolio/api/decorators/has-permission.decorator';
 import { HasPermissionGuard } from '@ghostfolio/api/guards/has-permission.guard';
+import { getInterval } from '@ghostfolio/api/helper/portfolio.helper';
 import { TransformDataSourceInRequestInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-request.interceptor';
 import { TransformDataSourceInResponseInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-response.interceptor';
 import type {
@@ -8,7 +9,7 @@ import type {
   UniqueAsset
 } from '@ghostfolio/common/interfaces';
 import { permissions } from '@ghostfolio/common/permissions';
-import type { RequestWithUser } from '@ghostfolio/common/types';
+import type { DateRange, RequestWithUser } from '@ghostfolio/common/types';
 
 import {
   Body,
@@ -19,6 +20,7 @@ import {
   Inject,
   Param,
   Post,
+  Query,
   UseGuards,
   UseInterceptors
 } from '@nestjs/common';
@@ -106,13 +108,18 @@ export class BenchmarkController {
   public async getBenchmarkMarketDataBySymbol(
     @Param('dataSource') dataSource: DataSource,
     @Param('startDateString') startDateString: string,
-    @Param('symbol') symbol: string
+    @Param('symbol') symbol: string,
+    @Query('range') dateRange: DateRange = 'max'
   ): Promise<BenchmarkMarketDataDetails> {
-    const startDate = new Date(startDateString);
+    const { endDate, startDate } = getInterval(
+      dateRange,
+      new Date(startDateString)
+    );
     const userCurrency = this.request.user.Settings.settings.baseCurrency;
 
     return this.benchmarkService.getMarketDataBySymbol({
       dataSource,
+      endDate,
       startDate,
       symbol,
       userCurrency