+++ apps/api/src/app/portfolio/portfolio.controller.ts
@@ -6,6 +6,7 @@ import {
   hasNotDefinedValuesInObject,
   nullifyValuesInObject
 } from '@ghostfolio/api/helper/object.helper';
+import { getInterval } from '@ghostfolio/api/helper/portfolio.helper';
 import { RedactValuesInResponseInterceptor } from '@ghostfolio/api/interceptors/redact-values-in-response.interceptor';
 import { TransformDataSourceInRequestInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-request.interceptor';
 import { TransformDataSourceInResponseInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-response.interceptor';
@@ -236,8 +237,12 @@ export class PortfolioController {
       await this.impersonationService.validateImpersonationId(impersonationId);
     const userCurrency = this.request.user.Settings.settings.baseCurrency;
 
+    const { endDate, startDate } = getInterval(dateRange);
+
     const { activities } = await this.orderService.getOrders({
+      endDate,
       filters,
+      startDate,
       userCurrency,
       userId: impersonationUserId || this.request.user.id,
       types: ['DIVIDEND']
@@ -245,7 +250,6 @@ export class PortfolioController {
 
     let dividends = await this.portfolioService.getDividends({
       activities,
-      dateRange,
       groupBy
     });