+++ apps/client/src/app/pages/portfolio/analysis/analysis-page.component.ts
@@ -352,6 +352,7 @@ export class AnalysisPageComponent implements OnDestroy, OnInit {
           .fetchBenchmarkBySymbol({
             dataSource,
             symbol,
+            range: this.user?.settings?.dateRange,
             startDate: this.firstOrderDate
           })
           .pipe(takeUntil(this.unsubscribeSubject))